move `Configuration` into the root module
Otherwise we would need `pub(restricted)` later on.
